+++ src/PrintOrder/PrintOrder.js
@@ -3,12 +3,16 @@ import PropTypes from 'prop-types';
 import { useReactToPrint } from 'react-to-print';
 import { useIntl } from 'react-intl';
 
-import { stripesConnect, useOkapiKy } from '@folio/stripes/core';
+import {
+  stripesConnect,
+  useOkapiKy,
+  useStripes,
+} from '@folio/stripes/core';
 
 import { exportManifest, getExportData } from '../common/ExportSettingsModal/utils';
 
-import PrintContent from './PrintContent';
 import { hydrateOrderToPrint } from './hydrateOrderToPrint';
+import PrintContent from './PrintContent';
 import {
   getOrderPrintData,
   getPrintPageStyles,
@@ -17,9 +21,11 @@ import {
 export const PrintOrderComponent = ({ mutator, order, orderLine, onCancel }) => {
   const intl = useIntl();
   const ky = useOkapiKy();
+  const stripes = useStripes();
 
   const [printableOrder, setPrintableOrder] = useState();
 
+  const systemCurrency = stripes.currency;
   const componentRef = useRef();
   const handlePrint = useReactToPrint({
     pageStyle: getPrintPageStyles(),
@@ -35,7 +41,9 @@ export const PrintOrderComponent = ({ mutator, order, orderLine, onCancel }) => 
         ? { lines: await getExportData(mutator, linesToPrint, [order], intl) }
         : await getOrderPrintData(ky, order);
 
-      setPrintableOrder(hydrateOrderToPrint({
+      setPrintableOrder(await hydrateOrderToPrint({
+        ky,
+        systemCurrency,
         order: {
           ...order,
           ...printData,